# ~3x test execution speedup from updating to `jest@30` and moving from `ts-jest` to `@swc/jest`