--- a/jest.config.js
+++ b/jest.config.js
@@ -9,7 +9,7 @@ export default {
     "/node_modules/(?!eventsource)/"
   ],
   testPathIgnorePatterns: ["/node_modules/", "/dist/"],
-    transform: {
+  transform: {
     '^.+\\.(t|j)sx?$': '@swc/jest',
   },
 };