--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@modelcontextprotocol/sdk",
-  "version": "1.17.5",
+  "version": "1.17.1",
   "description": "Model Context Protocol implementation for TypeScript",
   "license": "MIT",
   "author": "Anthropic, PBC (https://anthropic.com)",
@@ -19,18 +19,6 @@
     "mcp"
   ],
   "exports": {
-    ".": {
-      "import": "./dist/esm/index.js",
-      "require": "./dist/cjs/index.js"
-    },
-    "./client": {
-      "import": "./dist/esm/client/index.js",
-      "require": "./dist/cjs/client/index.js"
-    },
-    "./server": {
-      "import": "./dist/esm/server/index.js",
-      "require": "./dist/cjs/server/index.js"
-    },
     "./*": {
       "import": "./dist/esm/*",
       "require": "./dist/cjs/*"
@@ -90,7 +78,7 @@
     "@types/supertest": "^6.0.2",
     "@types/ws": "^8.5.12",
     "eslint": "^9.8.0",
-    "jest": "^30.0.5",
+    "jest": "^30.1.3",
     "supertest": "^7.0.0",
     "tsx": "^4.16.5",
     "typescript": "^5.5.4",