Jest v30 has performance improvements. However, while `ts-jest` claims it works with `jest@30` as a peer dependency, it requires a number of changes to the project tsconfig to get it working. Whereas, switching to `@swc/jest` works with `jest@29` and `jest@30` and it's ~2x as fast as `ts-jest` even before updating to `jest@30`.

### Current test execution time:
```
Time:        3.454 s, estimated 6 s
```

### Moving to `@swc/jest`:
```
Time:        1.405 s
```

### Moving to `@swc/jest` and updating to `jest@30`:
```
Time:        1.212 s
```

## Motivation and Context
Much faster test execution

## How Has This Been Tested?
Running the test suite, confirming the same number of tests ran, the same number passed, and the execution time was significantly less.

## Breaking Changes
None.

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation update

## Checklist
- [x] I have read the [MCP Documentation](https://modelcontextprotocol.io)
- [x] My code follows the repository's style guidelines
- [x] New and existing tests pass locally
- [x] I have added appropriate error handling
- [x] I have added or updated documentation as needed

## Additional context