@KKonstantinov 2025-10-06
Would like to compare this to and consider https://github.com/vitest-dev/vitest (which is very easy to switch to from jest) while we're at it. Jest appears to be very slow to be updated against new Node features in the last years, which is also kind of shown by the need of this PR in the first place.