@jake-danton 2025-10-07
> Would like to compare this to and consider https://github.com/vitest-dev/vitest (which is very easy to switch to from jest) while we're at it. Jest appears to be very slow to be updated against new Node features in the last years, which is also kind of shown by the need of this PR in the first place.

I am happy to do that as I already migrated the [@enth/mcp-sdk](https://github.com/enthspace/mcp-sdk) fork over to Vitest for the added speed and built in benchmarks and code coverage.