@jake-danton 2025-10-07
@KKonstantinov I think we should also discuss moving tests into a dedicated `test` folder instead of in `src` and at the same time, move `examples` (which are currently exported as part of the package), `integration-tests`, `__mocks__`, and anything else not meant for export.